Add support to build using cmake config files installed by libzip
Add support to build using cmake config files installed by libzip.
Add and option that controls if config mode is to be used,
this is on per default on non windows systems.
The package name used by libzip is libzip not LIBZIP so add
a variable that holds this and use it where needed in
CMakeLists.txt and Config.cmake.in.
FindLIBZIP.cmake is unneeded in config mode so don't install
it if config mode is used.
This commit also includes changes due to line endings.
This the editor that saves in \n instead of \r\n.