Support using GNUInstallDirs for installation
Add an option that enables use of GNUInstallDirs.
Set its default value to ON on unix-like systems.
When the option is off files are installed into
the same locations as before.
Remove the DESTINATION arguments from install,
these have default values that are the same as
the ones used with or without GNUIinstallDirs
then using cmake 3.14 and above, up the least
required cmake version to 3.14 due to this.
Install cmake config files into libdir/cmake
when using GNUInstalldirs instead of
share/libzipp since libzippp is not a header
only library.