Add pkg-config file generation
Add pkg-config file generation.
In order to generate the right
values in the pkg-config file
GNUInstallDirs has to be used.
The pkg-config file in only
relevant for development so
only install it when the header
is installed, it also uses the
name of the shared lib for
linking so only install it
when the shared lib is built.