Fix errors in FindLIBZIP.cmake
FreeBSD native packages put
pkg-config files in libdata/pkgconfig
so add it to PATH_SUFFIXES.
Some find_package calls have the
imported targets as argument where
the module name is to be used, fix
this.
Change lgnutls to -lnettle since this
is probably what is intended since
the statements below it concerns the
nettle library.
Bump the required version of cmake to
version 3.16 since the GnuTLS::GnuTLS
imported target is only available in
cmake 3.16 and later.